Update install script

+++ install.sh
@@ -2,7 +2,10 @@
 
 GITHUB="https://github.com/frousselet/dotfiles.git"
 
+
+###########
 # OSX: Check Homebrew installation
+######
 if [[ "$OSTYPE" == "darwin"* ]]; then
     if [[ ! command -v brew > /dev/null ]]; then
         printf "Installing Homebrew...\n"
@@ -17,7 +20,10 @@ if [[ "$OSTYPE" == "darwin"* ]]; then
     fi
 fi
 
+
+###########
 # ALL: Check Git installation
+######
 if [[ ! command -v git > /dev/null ]]; then
     printf "Installing Git...\n"
     if [[ "$OSTYPE" == "darwin"* ]]; then
@@ -32,20 +38,32 @@ if [[ ! command -v git > /dev/null ]]; then
     fi
 fi
 
+
+###########
 # ALL: Clone dotfiles
+######
 rm -rf "$HOME/.dotfiles"
 git clone "$GITHUB" "$HOME/.dotfiles"
 
-# ALL: Install Oh-My-ZSH
-if command -v curl > /dev/null; then
-    sh -c "$(curl -fsSL https://raw.github.com/ohmyzsh/ohmyzsh/master/tools/install.sh)"
-elif command -v wget > /dev/null; then
-    sh -c "$(wget https://raw.github.com/ohmyzsh/ohmyzsh/master/tools/install.sh -O -)"
-else
-    printf "curl or wget not installed!\n"
-    exit 1
-fi
 
-# ALL: Enable ZSH configuration
-ln -sf "$HOME/.dotfiles/zsh/.zshrc" "$HOME/.zshrc"
-. "$HOME/.zshrc"
+###########
+# ALL: Configure ZSH
+######
+printf "Configure ZSH? (Y/n) "
+read CONFIGURE_ZSH
+
+if [[ "$CONFIGURE_ZSH" == "y" || "$CONFIGURE_ZSH" == "yes" ]];then
+    printf "Configuring ZSH...\n"
+    if command -v curl > /dev/null; then
+        sh -c "$(curl -fsSL https://raw.github.com/ohmyzsh/ohmyzsh/master/tools/install.sh)"
+    elif command -v wget > /dev/null; then
+        sh -c "$(wget https://raw.github.com/ohmyzsh/ohmyzsh/master/tools/install.sh -O -)"
+    else
+        printf "curl or wget not installed!\n"
+        exit 1
+    fi
+    ln -sf "$HOME/.dotfiles/zsh/.zshrc" "$HOME/.zshrc"
+    . "$HOME/.zshrc"
+else
+    printf "ZSH configuration skipped.\n"
+fi
